+++ app/UseCases/Event/EventCommunityUseCase.php
@@ -46,31 +46,6 @@ class EventCommunityUseCase
     }
 
     /**
-     * 使用可能な絵文字の一覧を取得する
-     *
-     * @return array 絵文字の配列
-     */
-    public function getEmojis()
-    {
-        return [
-            'face_and_persons' =>     [
-                '🙂', '😀', '😃', '😄', '😁', '😅', '😆', '🤣', '😂', '🙃', '😉', '😊', '😇', '😎', '🧐',
-                '🥳', '🥰', '😍', '🤩', '😙', '🥲', '😜', '🤪', '😝', '🤑', '🤗',
-                '😒', '🙄', '😬', '😮‍💨', '🤥', '😪', '😴', '😌', '😔', '🤤', '😷', '🤒', '🤕', '🤢', '🤮', '🤧', '🥵', '🥶', '🥴', '😵',
-                '🤯', '😕', '😟', '🙁', '☹', '😮', '😯', '😲', '😳', '🥺', '😦', '😧', '😨', '😰', '😥', '😢', '😭', '😱', '😖', '😣',
-                '😞', '😓', '😩', '😫', '🥱', '😤', '😡', '😠', '🤬',  '💩',
-
-            ],
-            'emotions' =>     ['💖', '💗', '💕', '❣', '💔', '💯', '💢', '💥', '💫', '💦', '💬', '💤'],
-            'tasks' => [
-                '✅', '❌', '🎉', '👀', '🙏', '👍', '🙎', '🙎‍♂‍', '🙎‍♀‍', '🙅', '🙅‍♂‍', '🙅‍♀‍', '🙆', '🙆‍♂‍', '🙆‍♀‍', '💁', '💁‍♂‍',
-                '💁‍♀‍', '🙋', '🙋‍♂‍', '🙋‍♀‍', '🙇', '🙇‍♂‍', '🙇‍♀‍', '🤷', '🤷‍♂‍', '🤷‍♀‍'
-            ]
-        ];
-    }
-
-
-    /**
      * イベントへのアクセス権限をチェックする
      *
      * @param int $id イベントのID
