+++ lang/ja/validation.php
@@ -270,5 +270,6 @@ return [
         'number_of_recruits' => '人数',
         'image_path' => '画像',
         "login_id" => "ログインID",
+        "emoji" => "絵文字",
     ]
 ];
